tests/wpt/web-platform-tests/websockets/constructor/007.html

<!doctype html>
<title>WebSockets: new WebSocket(url, null char)</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../constants.js?pipe=sub></script>
<meta name="variant" content="">
<meta name="variant" content="?wss">
<div id=log></div>
<script>
test(function() {
  assert_throws("SyntaxError", function() {
    new WebSocket(SCHEME_DOMAIN_PORT + '/empty-message',
                  'a' + String.fromCharCode(0) + 'b')
  })
});
</script>
